This dataset is used as Final Project in Week 8
In week 8, students will use the skills gained from the AutoAI exercises in Week 6 to complete the same steps, yet using a different dataset. This dataset is used in Kaggle competitions and numerous hackathons. It is about uncovering a bank's lending practices and potential bias built into them by students.